Api Dokcling/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
docling==0.1.0
python-dotenv==1.0.0
